Test access and share file mode for Windows C API
This commit attempts to set the most permissive share and access modes
when opening or creating the standard input, output, and error files
for the C API on Windows.
This attempts to fix a bug on Windows where the standard error file
seems to continue to have a lock when tried to access from the host.

Signed-off-by: Radu M <root@radu.sh>